myporject/requirements.txt
pandas

django==2.1.5
googlemaps

sqlalchemy

pymysql

matplotlib

confluent_kafka

requests

bs4

lxml

opencc-python-reimplemented

jieba

sklearn 

joblib

numpy

bs4

textrank4zh

snownlp

shortcuts

transformers==2.1.1

wordcloud

elasticsearch

argparse

django-simple-captcha

